+++ Cargo.toml
@@ -90,6 +90,7 @@ kailua-common = { path = "crates/common" }
 kailua-contracts = { path = "crates/contracts" }
 kailua-host = { path = "bin/host" }
 kailua-proposer = { path = "crates/proposer" }
+kailua-prover = { path = "crates/prover" }
 kailua-sync = { path = "crates/sync" }
 
 # Kona git = "https://github.com/op-rs/kona", tag = "kona-client/v1.0.1"
