+++ bin/client/Cargo.toml
@@ -31,6 +31,7 @@ alloy-primitives = { workspace = true, features = ["map-hashbrown"] }
 kailua-build.workspace = true
 kailua-common.workspace = true
 kailua-sync.workspace = true
+kailua-prover.workspace = true
 
 kona-cli.workspace = true
 kona-derive.workspace = true
