+++ bin/host/Cargo.toml
@@ -30,6 +30,7 @@ alloy-eips.workspace = true
 kailua-build.workspace = true
 kailua-client.workspace = true
 kailua-common.workspace = true
+kailua-prover.workspace = true
 kailua-sync.workspace = true
 
 # Kona
