+++ bin/host/src/main.rs
@@ -17,7 +17,6 @@ use alloy_eips::BlockNumberOrTag;
 use alloy_primitives::B256;
 use anyhow::{anyhow, bail, Context};
 use clap::Parser;
-use kailua_client::proving::ProvingError;
 use kailua_common::boot::StitchedBootInfo;
 use kailua_host::args::KailuaHostArgs;
 use kailua_host::channel::AsyncChannel;
@@ -25,6 +24,7 @@ use kailua_host::config::generate_rollup_config;
 use kailua_host::preflight::{concurrent_execution_preflight, fetch_precondition_data};
 use kailua_host::server::create_disk_kv_store;
 use kailua_host::tasks::{handle_oneshot_tasks, Cached, Oneshot, OneshotResult};
+use kailua_prover::ProvingError;
 use kailua_sync::provider::optimism::OpNodeProvider;
 use std::collections::BinaryHeap;
 use std::env::set_var;
