+++ bin/host/src/preflight.rs
@@ -20,9 +20,9 @@ use alloy_eips::eip4844::IndexedBlobHash;
 use alloy_eips::BlockNumberOrTag;
 use alloy_primitives::B256;
 use anyhow::bail;
-use kailua_client::proving::ProvingError;
 use kailua_common::blobs::BlobFetchRequest;
 use kailua_common::precondition::PreconditionValidationData;
+use kailua_prover::ProvingError;
 use kailua_sync::provider::optimism::OpNodeProvider;
 use kona_genesis::RollupConfig;
 use kona_preimage::{PreimageKey, PreimageKeyType};
