+++ crates/proposer/Cargo.toml
@@ -12,31 +12,23 @@ categories.workspace = true
 
 [dependencies]
 anyhow.workspace = true
-async-channel.workspace = true
-bytemuck.workspace = true
 clap.workspace = true
-hex.workspace = true
-tempfile.workspace = true
 tokio.workspace = true
 tokio-retry.workspace = true
-tracing-subscriber.workspace = true
 tracing.workspace = true
 
 opentelemetry.workspace = true
 
 alloy = { workspace = true, features = ["full", "kzg", "signer-aws", "signer-gcp"] }
 
-kailua-build.workspace = true
 kailua-common.workspace = true
 kailua-contracts.workspace = true
 kailua-sync.workspace = true
 
-kona-cli.workspace = true
-kona-protocol.workspace = true
-
-risc0-zkvm.workspace = true
-
 [features]
 devnet = [
     "kailua-sync/devnet"
+]
+rebuild-fpvm = [
+    "kailua-sync/rebuild-fpvm"
 ]
