+++ crates/prover/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "kailua-cli"
+name = "kailua-prover"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -12,42 +12,40 @@ categories.workspace = true
 
 [dependencies]
 anyhow.workspace = true
-async-channel.workspace = true
+async-trait.workspace = true
+bincode.workspace = true
 bytemuck.workspace = true
+c-kzg.workspace = true
 clap.workspace = true
 hex.workspace = true
-tempfile.workspace = true
-tokio.workspace = true
-tokio-retry.workspace = true
-tracing-subscriber.workspace = true
+human_bytes.workspace = true
+rkyv.workspace = true
+thiserror.workspace = true
 tracing.workspace = true
+tracing-subscriber.workspace = true
+tokio.workspace = true
 
-opentelemetry.workspace = true
-
-alloy = { workspace = true, features = ["full", "kzg", "signer-aws", "signer-gcp"] }
+alloy = { workspace = true, features = ["full", "kzg"] }
+alloy-primitives = { workspace = true, features = ["map-hashbrown"] }
 
 kailua-build.workspace = true
-kailua-client.workspace = true
 kailua-common.workspace = true
-kailua-contracts.workspace = true
-kailua-host.workspace = true
-kailua-proposer.workspace = true
-kailua-prover.workspace = true
 kailua-sync.workspace = true
 
 kona-cli.workspace = true
+kona-derive.workspace = true
+kona-preimage.workspace = true
+kona-proof.workspace = true
 kona-protocol.workspace = true
+kona-std-fpvm.workspace = true
 
+bonsai-sdk.workspace = true
+boundless-market.workspace = true
+
+risc0-ethereum-contracts.workspace = true
 risc0-zkvm.workspace = true
 
-[dev-dependencies]
-assert_cmd = "2.0.16"
-
 [features]
-devnet = [
-    "kailua-proposer/devnet",
-    "kailua-sync/devnet",
-]
 prove = [
     "risc0-zkvm/prove"
 ]
@@ -61,5 +59,6 @@ disable-dev-mode = [
     "risc0-zkvm/disable-dev-mode"
 ]
 rebuild-fpvm = [
-    "kailua-build/rebuild-fpvm"
+    "kailua-build/rebuild-fpvm",
+    "kailua-sync/rebuild-fpvm"
 ]
