+++ crates/prover/src/backends/bonsai.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::proving::{KailuaProveInfo, KailuaSessionStats};
-use crate::proving::{ProvingArgs, ProvingError};
+use crate::backends::{KailuaProveInfo, KailuaSessionStats};
+use crate::{ProvingArgs, ProvingError};
 use anyhow::{anyhow, Context};
 use bonsai_sdk::non_blocking::Client;
 use human_bytes::human_bytes;
