+++ crates/sync/Cargo.toml
@@ -46,3 +46,6 @@ kona-genesis.workspace = true
 
 [features]
 devnet = []
+rebuild-fpvm = [
+    "kailua-build/rebuild-fpvm"
+]
