+++ crates/sync/src/args.rs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 use crate::provider::ProviderArgs;
-use alloy::primitives::Address;
+use alloy::primitives::{Address, B256};
 use std::path::PathBuf;
 use std::str::FromStr;
 
@@ -33,4 +33,8 @@ pub struct SyncArgs {
 
 pub fn parse_address(s: &str) -> Result<Address, String> {
     Address::from_str(s).map_err(|_| format!("Invalid Address value: {}", s))
+}
+
+pub fn parse_b256(s: &str) -> Result<B256, String> {
+    B256::from_str(s).map_err(|_| format!("Invalid B256 value: {}", s))
 }
